refactor: decompose route comparison
